netlink: Fix signed integer comparison of len variable

+++ ell/netlink.c
@@ -272,7 +272,7 @@ static void can_read_data(struct l_io *io, void *user_data)
 		group = pktinfo->group;
 	}
 
-	for (nlmsg = iov.iov_base; NLMSG_OK(nlmsg, len);
+	for (nlmsg = iov.iov_base; NLMSG_OK(nlmsg, (uint32_t) len);
 					nlmsg = NLMSG_NEXT(nlmsg, len)) {
 		if (group > 0 && nlmsg->nlmsg_seq == 0) {
 			process_broadcast(netlink, group, nlmsg);
